フォロー/フォロー解除機能のCSRF対策追記

+++ src/main/java/actions/FollowAction.java
@@ -60,6 +60,8 @@ public class FollowAction extends ActionBase {
 		//		全ての従業員データの件数を取得
 		long employeeCount = service.countAll();
 
+		putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
+
 		putSessionScope(AttributeConst.FOLLOWS, follows); // 取得したフォローされる従業員データ
 
 		putRequestScope(AttributeConst.EMPLOYEES, employees); // 取得した従業員データ
@@ -80,51 +82,57 @@ public class FollowAction extends ActionBase {
 
 	//	フォローを行う
 	public void follow() throws ServletException, IOException {
-		//		セッションからログイン中の従業員情報を取得　
-		EmployeeView ev = (EmployeeView) getSessionScope(AttributeConst.LOGIN_EMP);
-		//		idを条件にフォローされる従業員情報を取得
-		EmployeeView follower = employeeService.findOne(toNumber(getRequestParam(AttributeConst.FOL_ID)));
+		if (checkToken()) {
+			//		セッションからログイン中の従業員情報を取得　
+			EmployeeView ev = (EmployeeView) getSessionScope(AttributeConst.LOGIN_EMP);
+			//		idを条件にフォローされる従業員情報を取得
+			EmployeeView follower = employeeService.findOne(toNumber(getRequestParam(AttributeConst.FOL_ID)));
 
-		//		パラメータの値を元にフォローされる従業員情報のインスタンスを作成する
-		FollowView fv = new FollowView(
-				null, // Integer id
-				ev, // Employee employee_id
-				follower, // Employee follow_id
-				null, // LocalDateTime created_at
-				null, // LocalDateTime updated_at
-				AttributeConst.DEL_FLAG_FALSE.getIntegerValue() // Integer delete_flag
-		);
-		//		登録
-		service.create(fv);
+			//		パラメータの値を元にフォローされる従業員情報のインスタンスを作成する
+			FollowView fv = new FollowView(
+					null, // Integer id
+					ev, // Employee employee_id
+					follower, // Employee follow_id
+					null, // LocalDateTime created_at
+					null, // LocalDateTime updated_at
+					AttributeConst.DEL_FLAG_FALSE.getIntegerValue() // Integer delete_flag
+			);
+			//		登録
+			service.create(fv);
 
-		//		フラッシュメッセージをセッションに登録
-		putSessionScope(AttributeConst.FLUSH, fv.getFollow_id().getName() + MessageConst.I_FOLLOWED.getMessage());
+			//		フラッシュメッセージをセッションに登録
+			putSessionScope(AttributeConst.FLUSH, fv.getFollow_id().getName() + MessageConst.I_FOLLOWED.getMessage());
 
-		redirect(ForwardConst.ACT_FOL, ForwardConst.CMD_INDEX);
+			redirect(ForwardConst.ACT_FOL, ForwardConst.CMD_INDEX);
+		}
 	}
 
 	public void destroy() throws ServletException, IOException {
-		//		ログイン中の従業員のフォローリストを取得
-		List<FollowView> follower = getSessionScope(AttributeConst.FOLLOWS);
+		if (checkToken()) {
+			//		ログイン中の従業員のフォローリストを取得
+			List<FollowView> follower = getSessionScope(AttributeConst.FOLLOWS);
 
-		//		フォローリストの中から フォローされる従業員id と送信されたリクエストパラメータ フォロー解除ボタンを押した解除対象の従業員id が同じ場合
-		//		destroyメソッドを実行
-		for (FollowView follow : follower) {
-			if (follow.getFollow_id().getId() == toNumber(getRequestParam(AttributeConst.FOL_ID))) {
-				//		論理削除メソッド
-				service.destroy(follow.getId());
-				//				フラッシュメッセージをセッションに登録
-				putSessionScope(AttributeConst.FLUSH,
-						follow.getFollow_id().getName() + MessageConst.I_UNFOLLOWED.getMessage());
-				break;
+			//		フォローリストの中から フォローされる従業員id と送信されたリクエストパラメータ フォロー解除ボタンを押した解除対象の従業員id が同じ場合
+			//		destroyメソッドを実行
+			for (FollowView follow : follower) {
+				if (follow.getFollow_id().getId() == toNumber(getRequestParam(AttributeConst.FOL_ID))) {
+					//		論理削除メソッド
+					service.destroy(follow.getId());
+					//				フラッシュメッセージをセッションに登録
+					putSessionScope(AttributeConst.FLUSH,
+							follow.getFollow_id().getName() + MessageConst.I_UNFOLLOWED.getMessage());
+					break;
+				}
 			}
-		}
 
-		//		一覧画面にリダイレクト
-		redirect(ForwardConst.ACT_FOL, ForwardConst.CMD_INDEX);
+			//		一覧画面にリダイレクト
+			redirect(ForwardConst.ACT_FOL, ForwardConst.CMD_INDEX);
+		}
 	}
 
+	//	従業員一覧を表示
 	public void showFollow() throws ServletException, IOException {
+
 		//		セッションからログイン中の従業員を取得
 		EmployeeView ev = (EmployeeView) getSessionScope(AttributeConst.LOGIN_EMP);
 
@@ -136,6 +144,9 @@ public class FollowAction extends ActionBase {
 
 		//全日報データの件数を取得
 		long reportsCount = reports.size();
+
+		putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
+
 		putRequestScope(AttributeConst.REP_FOL, reports); // 取得したフォロー中の従業員の日報データ
 		putRequestScope(AttributeConst.REP_FOL_COUNT, reportsCount); // フォロー中の従業員全ての日報データの件数
 		putRequestScope(AttributeConst.PAGE, page); //ページ数
